+++ nengo/__init__.py
@@ -48,9 +48,13 @@ from .neurons import (
     Izhikevich,
     LIF,
     LIFRate,
+    PoissonSpiking,
     RectifiedLinear,
+    RegularSpiking,
     Sigmoid,
     SpikingRectifiedLinear,
+    StochasticSpiking,
+    Tanh,
 )
 from .network import Network
 from .learning_rules import PES, BCM, Oja, Voja